Add `encoded_residuals` parser
Note: The remainder bit length can differ by (log2(parameter) + 1).pow()
according to the reference implementation. Testing it on an actually FLAC
file on that algorithm incorrectly parsed the file. This current
implementation works on that file and I will use it until I find a file that
messes with it.