Add trait for sample
This is mainly to kick off the fixing of decoding FLAC file with samples
larger than 16 bits. However, there is something to be said about only
allocating what is necessary depending on the sample size. I stuck with i32
because most files are 16 bits per sample, but that is short sighted and I
do want to actually support files 17-32 bits per sample.